pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "nyc-taxi-fare-mlops"
version = "1.0.0"
description = "Production-grade MLOps pipeline for NYC Taxi Fare Prediction"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0,<3.0",
    "numpy>=1.24,<2.0",
    "scikit-learn>=1.3,<2.0",
    "xgboost>=2.0,<3.0",
    "fastapi>=0.104,<1.0",
    "uvicorn[standard]>=0.24,<1.0",
    "prefect>=2.14,<3.0",
    "mlflow>=2.9,<3.0",
    "boto3>=1.29,<2.0",
    "psycopg2-binary>=2.9,<3.0",
    "prometheus-client>=0.19,<1.0",
    "streamlit>=1.29,<2.0",
    "httpx>=0.25,<1.0",
    "pyarrow>=14.0,<16.0",
    "pydantic>=2.5,<3.0",
    "python-dotenv>=1.0,<2.0",
    "plotly>=5.18,<6.0",
    "requests>=2.31,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4,<9.0",
    "pytest-cov>=4.1,<6.0",
    "pytest-asyncio>=0.23,<1.0",
    "ruff>=0.1.6,<1.0",
    "pre-commit>=3.5,<4.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "api*", "pipelines*", "monitoring*", "dashboard*"]

[tool.ruff]
target-version = "py310"
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"
